databricks-insight-agent/requirements.txt
databricks-sdk>=0.20.0
dspy-ai>=2.4.0
mlflow>=2.10.0
pandas>=2.0.0
pyspark>=3.5.0
openai>=1.0.0
python-dotenv>=1.0.0
structlog>=23.0.0
pydantic>=2.0.0
asyncio-mqtt>=0.16.0
sqlalchemy>=2.0.0
alembic>=1.12.0
fastapi>=0.104.0
uvicorn>=0.24.0